move EoT module to utilities
